requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.28.62
botocore==1.31.62
bravado==11.0.3
bravado-core==6.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
croniter==1.4.1
cvxopt==1.3.2
cycler==0.12.0
dateutils==0.6.12
deepdiff==6.6.0
deprecation==2.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
graphviz @ file:///var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_adtn416tfr/croots/recipe/python-graphviz_1660063186502/work
grpcio==1.59.0
h11==0.14.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
intervaltree==3.1.0
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.0.9.post0
lightning-cloud==0.5.39
lightning-utilities==0.9.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
monai==0.9.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
neptune==1.8.1
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
ordered-set==4.1.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pm4py==2.7.8
portalocker==2.8.2
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
pyDeprecate==0.3.2
pydotplus==2.0.2
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
s3transfer==0.7.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
StringDist==1.0.9
swagger-spec-validator==3.0.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==1.13.1
torchaudio==0.13.1
torchdata==0.5.1
torchmetrics==1.2.0
torchtext==0.14.1
torchvision==0.14.1
torchviz==0.0.2
tqdm==4.66.1
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.17
uvicorn==0.23.2
wcwidth==0.2.8
webcolors==1.13
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
